fixed some stray fcloses ( thx @c00kiemon5ter ) , ./make.sh is a light make script :P

+++ opengl_acquisition_shared_library/opengl_depth_and_color_renderer/src/main.c
@@ -98,7 +98,9 @@ int getOpenGLDepth(short * depth , unsigned int x,unsigned int y,unsigned int wi
          for ( i =0 ; i < (width-x); i ++ )
             {
                if (zbuffer[(height-1-yp)*stride+i]>=farPlane-nearPlane)  { depth[yp*stride+i]=  (short) 0;  depth[yp*stride+i+1]=  (short) 0; } else
-                                                                         { depth[yp*stride+i]=  65535 - zbuffer[(height-1-yp)*stride+i] * multiplier;
+                                                                         {
+
+                                                                           depth[yp*stride+i]=  (short) 65535 - zbuffer[(height-1-yp)*stride+i] * multiplier;
                                                                            //SOMETHING CRAZY IS HAPPENING HERE :
                                                                            //TODO :
                                                                            //if (depth[yp*stride+i]<=1) { depth[yp*stride+i]=0; }
@@ -110,8 +112,8 @@ int getOpenGLDepth(short * depth , unsigned int x,unsigned int y,unsigned int wi
     int i=0;
     for ( i =0 ; i < (width-x)*(height-y); i ++ )
       {
-        if (zbuffer[i]>=farPlane-nearPlane)  { depth[i]=   (short) 0; depth[i+1]=   (short) 0; } else
-                                             { depth[i]=  65535 - zbuffer[i] * multiplier; }
+        if (zbuffer[i]>=farPlane-nearPlane)  { depth[i]=   (unsigned short) 0; depth[i+1]=   (short) 0; } else
+                                             { depth[i]=   (unsigned short) 65535 - zbuffer[i] * multiplier; }
       }
     #endif
